Hermes project structure refactoring
