+++ src/main/java/agh/edu/hermes/checker/SlaViolationChecker.java
@@ -1,7 +1,7 @@
 package agh.edu.hermes.checker;
 
 
-import agh.edu.hermes.types.PolicyRule;
+import agh.edu.hermes.persistance.entities.PolicyRule;
 
 import java.util.List;
