+++ src/main/java/agh/edu/hermes/enums/RelationType.java
@@ -1,4 +1,4 @@
-package agh.edu.hermes.types.attributes;
+package agh.edu.hermes.enums;
 
 import java.util.List;
