+++ src/main/java/agh/edu/hermes/enums/SlaType.java
@@ -1,4 +1,4 @@
-package agh.edu.hermes.types.attributes;
+package agh.edu.hermes.enums;
 
 public enum SlaType {
     IAAS,
