+++ src/main/java/agh/edu/hermes/enums/ValueType.java
@@ -1,4 +1,4 @@
-package agh.edu.hermes.types.attributes;
+package agh.edu.hermes.enums;
 
 public enum ValueType {
     AVAILABILITY,
