+++ src/main/java/agh/edu/hermes/persistance/repositories/ActionRepository.java
@@ -1,6 +1,6 @@
-package agh.edu.hermes.repositories;
+package agh.edu.hermes.persistance.repositories;
 
-import agh.edu.hermes.types.attributes.Action;
+import agh.edu.hermes.persistance.entities.Action;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ActionRepository extends JpaRepository<Action, Long> {}
