+++ src/main/java/agh/edu/hermes/persistance/repositories/ClauseRepository.java
@@ -1,6 +1,6 @@
-package agh.edu.hermes.repositories;
+package agh.edu.hermes.persistance.repositories;
 
-import agh.edu.hermes.types.attributes.Clause;
+import agh.edu.hermes.persistance.entities.Clause;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ClauseRepository extends JpaRepository<Clause, Long> {}
