+++ src/main/java/agh/edu/hermes/persistance/repositories/PoliciesRepository.java
@@ -1,11 +1,8 @@
-package agh.edu.hermes.repositories;
+package agh.edu.hermes.persistance.repositories;
 
-import agh.edu.hermes.types.Policies;
-import agh.edu.hermes.types.PolicyRule;
+import agh.edu.hermes.persistance.entities.Policies;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface PoliciesRepository extends JpaRepository<Policies, Long> {}
