+++ src/main/java/agh/edu/hermes/persistance/repositories/SlaRepository.java
@@ -1,6 +1,6 @@
-package agh.edu.hermes.repositories;
+package agh.edu.hermes.persistance.repositories;
 
-import agh.edu.hermes.types.Sla;
+import agh.edu.hermes.persistance.entities.Sla;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
