+++ src/main/java/agh/edu/hermes/services/parsers/SlaParserService.java
@@ -1,6 +1,6 @@
 package agh.edu.hermes.services.parsers;
 
-import agh.edu.hermes.types.Sla;
+import agh.edu.hermes.persistance.entities.Sla;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import org.springframework.stereotype.Service;
