+++ src/main/resources/application.properties
@@ -8,7 +8,7 @@ server.port=${HERMES_PORT:8080}
 spring.datasource.url=jdbc:mysql://localhost:3306/pandora_box_db
 spring.datasource.username=hermes
 spring.datasource.password=hermes
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=create
 
 #json serialization responses
 spring.jackson.serialization.fail-on-empty-beans=false
